Store negative timeline feedback
When user taps "Mark as incorrect" for a timeline event, actually store it in
`timeline_feedback` table. We use a Boolean flag to determine whether a given row
is "correct" (either an amended time, or a confirmed un-amended time) or "incorrect"
(the user didn't give a new time, only told us that our original time is wrong).